requirements.txt
# Core PDF Processing
PyMuPDF==1.23.26

# Word Document Generation
python-docx==1.1.0

# Web Framework
Flask==3.0.2
Flask-CORS==4.0.0

# Utilities
Werkzeug==3.0.1
